Data.Array.Accelerate.Control.Lens

Description

Documentation

module Control.Lens

liftLens :: (Functor f, Unlift box s, Unlift box t, Unlift box b, Lift box a) => ((a -> f b) -> s -> f t) -> (box (Plain a) -> f (box (Plain b))) -> box (Plain s) -> f (box (Plain t))

Lift a Lens into a lens on the equivalent Accelerate term.

The instances exported by this package are all defined in terms of this function, using the definitions from the base lens package.