Data.Array.Accelerate.Control.Lens.Each

Description

Documentation

module Control.Lens.Each

Orphan instances

(Arrays a, Arrays b) => Each (Acc (a, a)) (Acc (b, b)) (Acc a) (Acc b)

Instance details

Methods

each :: Traversal (Acc (a, a)) (Acc (b, b)) (Acc a) (Acc b)

(Arrays a, Arrays b) => Each (Acc (a, a, a)) (Acc (b, b, b)) (Acc a) (Acc b)

Instance details

Methods

each :: Traversal (Acc (a, a, a)) (Acc (b, b, b)) (Acc a) (Acc b)

(Arrays a, Arrays b) => Each (Acc (a, a, a, a)) (Acc (b, b, b, b)) (Acc a) (Acc b)

Instance details

Methods

each :: Traversal (Acc (a, a, a, a)) (Acc (b, b, b, b)) (Acc a) (Acc b)

(Arrays a, Arrays b) => Each (Acc (a, a, a, a, a)) (Acc (b, b, b, b, b)) (Acc a) (Acc b)

Instance details

Methods

each :: Traversal (Acc (a, a, a, a, a)) (Acc (b, b, b, b, b)) (Acc a) (Acc b)

(Arrays a, Arrays b) => Each (Acc (a, a, a, a, a, a)) (Acc (b, b, b, b, b, b)) (Acc a) (Acc b)

Instance details

Methods

each :: Traversal (Acc (a, a, a, a, a, a)) (Acc (b, b, b, b, b, b)) (Acc a) (Acc b)

(Arrays a, Arrays b) => Each (Acc (a, a, a, a, a, a, a)) (Acc (b, b, b, b, b, b, b)) (Acc a) (Acc b)

Instance details

Methods

each :: Traversal (Acc (a, a, a, a, a, a, a)) (Acc (b, b, b, b, b, b, b)) (Acc a) (Acc b)

(Arrays a, Arrays b) => Each (Acc (a, a, a, a, a, a, a, a)) (Acc (b, b, b, b, b, b, b, b)) (Acc a) (Acc b)

Instance details

Methods

each :: Traversal (Acc (a, a, a, a, a, a, a, a)) (Acc (b, b, b, b, b, b, b, b)) (Acc a) (Acc b)

(Arrays a, Arrays b) => Each (Acc (a, a, a, a, a, a, a, a, a)) (Acc (b, b, b, b, b, b, b, b, b)) (Acc a) (Acc b)

Instance details

Methods

each :: Traversal (Acc (a, a, a, a, a, a, a, a, a)) (Acc (b, b, b, b, b, b, b, b, b)) (Acc a) (Acc b)

(Elt a, Elt b) => Each (Exp (a, a)) (Exp (b, b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (a, a)) (Exp (b, b)) (Exp a) (Exp b)

(Elt a, Elt b) => Each (Exp (a, a, a)) (Exp (b, b, b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (a, a, a)) (Exp (b, b, b)) (Exp a) (Exp b)

(Elt a, Elt b) => Each (Exp (a, a, a, a)) (Exp (b, b, b, b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (a, a, a, a)) (Exp (b, b, b, b)) (Exp a) (Exp b)

(Elt a, Elt b) => Each (Exp (a, a, a, a, a)) (Exp (b, b, b, b, b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (a, a, a, a, a)) (Exp (b, b, b, b, b)) (Exp a) (Exp b)

(Elt a, Elt b) => Each (Exp (a, a, a, a, a, a)) (Exp (b, b, b, b, b, b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (a, a, a, a, a, a)) (Exp (b, b, b, b, b, b)) (Exp a) (Exp b)

(Elt a, Elt b) => Each (Exp (a, a, a, a, a, a, a)) (Exp (b, b, b, b, b, b, b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (a, a, a, a, a, a, a)) (Exp (b, b, b, b, b, b, b)) (Exp a) (Exp b)

(Elt a, Elt b) => Each (Exp (a, a, a, a, a, a, a, a)) (Exp (b, b, b, b, b, b, b, b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (a, a, a, a, a, a, a, a)) (Exp (b, b, b, b, b, b, b, b)) (Exp a) (Exp b)

(Elt a, Elt b) => Each (Exp (a, a, a, a, a, a, a, a, a)) (Exp (b, b, b, b, b, b, b, b, b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (a, a, a, a, a, a, a, a, a)) (Exp (b, b, b, b, b, b, b, b, b)) (Exp a) (Exp b)

(Elt a, Elt b, Elt (Complex a), Elt (Complex b)) => Each (Exp (Complex a)) (Exp (Complex b)) (Exp a) (Exp b)

Instance details

Methods

each :: Traversal (Exp (Complex a)) (Exp (Complex b)) (Exp a) (Exp b)