Data.Array.Accelerate.Control.Lens.Shape

Description

Documentation

module Control.Lens.Tuple

Orphan instances

(Elt sh, Elt a, Elt a') => Field1 (Exp (sh :. a)) (Exp (sh :. a')) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (sh :. a)) (Exp (sh :. a')) (Exp a) (Exp a')

(Elt sh, Elt a, Elt b, Elt b') => Field2 (Exp ((sh :. b) :. a)) (Exp ((sh :. b') :. a)) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp ((sh :. b) :. a)) (Exp ((sh :. b') :. a)) (Exp b) (Exp b')

(Elt sh, Elt a, Elt b, Elt c, Elt c') => Field3 (Exp (((sh :. c) :. b) :. a)) (Exp (((sh :. c') :. b) :. a)) (Exp c) (Exp c')

Instance details

Methods

_3 :: Lens (Exp (((sh :. c) :. b) :. a)) (Exp (((sh :. c') :. b) :. a)) (Exp c) (Exp c')

(Elt sh, Elt a, Elt b, Elt c, Elt d, Elt d') => Field4 (Exp ((((sh :. d) :. c) :. b) :. a)) (Exp ((((sh :. d') :. c) :. b) :. a)) (Exp d) (Exp d')

Instance details

Methods

_4 :: Lens (Exp ((((sh :. d) :. c) :. b) :. a)) (Exp ((((sh :. d') :. c) :. b) :. a)) (Exp d) (Exp d')

(Elt sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt e') => Field5 (Exp (((((sh :. e) :. d) :. c) :. b) :. a)) (Exp (((((sh :. e') :. d) :. c) :. b) :. a)) (Exp e) (Exp e')

Instance details

Methods

_5 :: Lens (Exp (((((sh :. e) :. d) :. c) :. b) :. a)) (Exp (((((sh :. e') :. d) :. c) :. b) :. a)) (Exp e) (Exp e')

(Elt sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt f') => Field6 (Exp ((((((sh :. f) :. e) :. d) :. c) :. b) :. a)) (Exp ((((((sh :. f') :. e) :. d) :. c) :. b) :. a)) (Exp f) (Exp f')

Instance details

Methods

_6 :: Lens (Exp ((((((sh :. f) :. e) :. d) :. c) :. b) :. a)) (Exp ((((((sh :. f') :. e) :. d) :. c) :. b) :. a)) (Exp f) (Exp f')

(Elt sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt g') => Field7 (Exp (((((((sh :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp (((((((sh :. g') :. f) :. e) :. d) :. c) :. b) :. a)) (Exp g) (Exp g')

Instance details

Methods

_7 :: Lens (Exp (((((((sh :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp (((((((sh :. g') :. f) :. e) :. d) :. c) :. b) :. a)) (Exp g) (Exp g')

(Elt sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt h') => Field8 (Exp ((((((((sh :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp ((((((((sh :. h') :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp h) (Exp h')

Instance details

Methods

_8 :: Lens (Exp ((((((((sh :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp ((((((((sh :. h') :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp h) (Exp h')

(Elt sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt i, Elt i') => Field9 (Exp (((((((((sh :. i) :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp (((((((((sh :. i') :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp i) (Exp i')

Instance details

Methods

_9 :: Lens (Exp (((((((((sh :. i) :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp (((((((((sh :. i') :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp i) (Exp i')