Data.Array.Accelerate.Control.Lens.Tuple

Description

Documentation

module Control.Lens.Tuple

Orphan instances

(Arrays a, Arrays a', Arrays b) => Field1 (Acc (a, b)) (Acc (a', b)) (Acc a) (Acc a')

Instance details

Methods

_1 :: Lens (Acc (a, b)) (Acc (a', b)) (Acc a) (Acc a')

(Arrays a, Arrays a', Arrays b, Arrays c) => Field1 (Acc (a, b, c)) (Acc (a', b, c)) (Acc a) (Acc a')

Instance details

Methods

_1 :: Lens (Acc (a, b, c)) (Acc (a', b, c)) (Acc a) (Acc a')

(Arrays a, Arrays a', Arrays b, Arrays c, Arrays d) => Field1 (Acc (a, b, c, d)) (Acc (a', b, c, d)) (Acc a) (Acc a')

Instance details

Methods

_1 :: Lens (Acc (a, b, c, d)) (Acc (a', b, c, d)) (Acc a) (Acc a')

(Arrays a, Arrays a', Arrays b, Arrays c, Arrays d, Arrays e) => Field1 (Acc (a, b, c, d, e)) (Acc (a', b, c, d, e)) (Acc a) (Acc a')

Instance details

Methods

_1 :: Lens (Acc (a, b, c, d, e)) (Acc (a', b, c, d, e)) (Acc a) (Acc a')

(Arrays a, Arrays a', Arrays b, Arrays c, Arrays d, Arrays e, Arrays f) => Field1 (Acc (a, b, c, d, e, f)) (Acc (a', b, c, d, e, f)) (Acc a) (Acc a')

Instance details

Methods

_1 :: Lens (Acc (a, b, c, d, e, f)) (Acc (a', b, c, d, e, f)) (Acc a) (Acc a')

(Arrays a, Arrays a', Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g) => Field1 (Acc (a, b, c, d, e, f, g)) (Acc (a', b, c, d, e, f, g)) (Acc a) (Acc a')

Instance details

Methods

_1 :: Lens (Acc (a, b, c, d, e, f, g)) (Acc (a', b, c, d, e, f, g)) (Acc a) (Acc a')

(Arrays a, Arrays a', Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays h) => Field1 (Acc (a, b, c, d, e, f, g, h)) (Acc (a', b, c, d, e, f, g, h)) (Acc a) (Acc a')

Instance details

Methods

_1 :: Lens (Acc (a, b, c, d, e, f, g, h)) (Acc (a', b, c, d, e, f, g, h)) (Acc a) (Acc a')

(Arrays a, Arrays a', Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays h, Arrays i) => Field1 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a', b, c, d, e, f, g, h, i)) (Acc a) (Acc a')

Instance details

Methods

_1 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a', b, c, d, e, f, g, h, i)) (Acc a) (Acc a')

(Elt a, Elt a', Elt b) => Field1 (Exp (a, b)) (Exp (a', b)) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (a, b)) (Exp (a', b)) (Exp a) (Exp a')

(Elt a, Elt a', Elt b, Elt c) => Field1 (Exp (a, b, c)) (Exp (a', b, c)) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (a, b, c)) (Exp (a', b, c)) (Exp a) (Exp a')

(Elt a, Elt a', Elt b, Elt c, Elt d) => Field1 (Exp (a, b, c, d)) (Exp (a', b, c, d)) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (a, b, c, d)) (Exp (a', b, c, d)) (Exp a) (Exp a')

(Elt a, Elt a', Elt b, Elt c, Elt d, Elt e) => Field1 (Exp (a, b, c, d, e)) (Exp (a', b, c, d, e)) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (a, b, c, d, e)) (Exp (a', b, c, d, e)) (Exp a) (Exp a')

(Elt a, Elt a', Elt b, Elt c, Elt d, Elt e, Elt f) => Field1 (Exp (a, b, c, d, e, f)) (Exp (a', b, c, d, e, f)) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (a, b, c, d, e, f)) (Exp (a', b, c, d, e, f)) (Exp a) (Exp a')

(Elt a, Elt a', Elt b, Elt c, Elt d, Elt e, Elt f, Elt g) => Field1 (Exp (a, b, c, d, e, f, g)) (Exp (a', b, c, d, e, f, g)) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (a, b, c, d, e, f, g)) (Exp (a', b, c, d, e, f, g)) (Exp a) (Exp a')

(Elt a, Elt a', Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h) => Field1 (Exp (a, b, c, d, e, f, g, h)) (Exp (a', b, c, d, e, f, g, h)) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (a, b, c, d, e, f, g, h)) (Exp (a', b, c, d, e, f, g, h)) (Exp a) (Exp a')

(Elt a, Elt a', Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt i) => Field1 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a', b, c, d, e, f, g, h, i)) (Exp a) (Exp a')

Instance details

Methods

_1 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a', b, c, d, e, f, g, h, i)) (Exp a) (Exp a')

(Arrays a, Arrays b, Arrays b') => Field2 (Acc (a, b)) (Acc (a, b')) (Acc b) (Acc b')

Instance details

Methods

_2 :: Lens (Acc (a, b)) (Acc (a, b')) (Acc b) (Acc b')

(Arrays a, Arrays b, Arrays b', Arrays c) => Field2 (Acc (a, b, c)) (Acc (a, b', c)) (Acc b) (Acc b')

Instance details

Methods

_2 :: Lens (Acc (a, b, c)) (Acc (a, b', c)) (Acc b) (Acc b')

(Arrays a, Arrays b, Arrays b', Arrays c, Arrays d) => Field2 (Acc (a, b, c, d)) (Acc (a, b', c, d)) (Acc b) (Acc b')

Instance details

Methods

_2 :: Lens (Acc (a, b, c, d)) (Acc (a, b', c, d)) (Acc b) (Acc b')

(Arrays a, Arrays b, Arrays b', Arrays c, Arrays d, Arrays e) => Field2 (Acc (a, b, c, d, e)) (Acc (a, b', c, d, e)) (Acc b) (Acc b')

Instance details

Methods

_2 :: Lens (Acc (a, b, c, d, e)) (Acc (a, b', c, d, e)) (Acc b) (Acc b')

(Arrays a, Arrays b, Arrays b', Arrays c, Arrays d, Arrays e, Arrays f) => Field2 (Acc (a, b, c, d, e, f)) (Acc (a, b', c, d, e, f)) (Acc b) (Acc b')

Instance details

Methods

_2 :: Lens (Acc (a, b, c, d, e, f)) (Acc (a, b', c, d, e, f)) (Acc b) (Acc b')

(Arrays a, Arrays b, Arrays b', Arrays c, Arrays d, Arrays e, Arrays f, Arrays g) => Field2 (Acc (a, b, c, d, e, f, g)) (Acc (a, b', c, d, e, f, g)) (Acc b) (Acc b')

Instance details

Methods

_2 :: Lens (Acc (a, b, c, d, e, f, g)) (Acc (a, b', c, d, e, f, g)) (Acc b) (Acc b')

(Arrays a, Arrays b, Arrays b', Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays h) => Field2 (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b', c, d, e, f, g, h)) (Acc b) (Acc b')

Instance details

Methods

_2 :: Lens (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b', c, d, e, f, g, h)) (Acc b) (Acc b')

(Arrays a, Arrays b, Arrays b', Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays h, Arrays i) => Field2 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b', c, d, e, f, g, h, i)) (Acc b) (Acc b')

Instance details

Methods

_2 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b', c, d, e, f, g, h, i)) (Acc b) (Acc b')

(Elt a, Elt b, Elt b') => Field2 (Exp (a, b)) (Exp (a, b')) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp (a, b)) (Exp (a, b')) (Exp b) (Exp b')

(Elt a, Elt b, Elt b', Elt c) => Field2 (Exp (a, b, c)) (Exp (a, b', c)) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp (a, b, c)) (Exp (a, b', c)) (Exp b) (Exp b')

(Elt a, Elt b, Elt b', Elt c, Elt d) => Field2 (Exp (a, b, c, d)) (Exp (a, b', c, d)) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp (a, b, c, d)) (Exp (a, b', c, d)) (Exp b) (Exp b')

(Elt a, Elt b, Elt b', Elt c, Elt d, Elt e) => Field2 (Exp (a, b, c, d, e)) (Exp (a, b', c, d, e)) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp (a, b, c, d, e)) (Exp (a, b', c, d, e)) (Exp b) (Exp b')

(Elt a, Elt b, Elt b', Elt c, Elt d, Elt e, Elt f) => Field2 (Exp (a, b, c, d, e, f)) (Exp (a, b', c, d, e, f)) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp (a, b, c, d, e, f)) (Exp (a, b', c, d, e, f)) (Exp b) (Exp b')

(Elt a, Elt b, Elt b', Elt c, Elt d, Elt e, Elt f, Elt g) => Field2 (Exp (a, b, c, d, e, f, g)) (Exp (a, b', c, d, e, f, g)) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp (a, b, c, d, e, f, g)) (Exp (a, b', c, d, e, f, g)) (Exp b) (Exp b')

(Elt a, Elt b, Elt b', Elt c, Elt d, Elt e, Elt f, Elt g, Elt h) => Field2 (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b', c, d, e, f, g, h)) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b', c, d, e, f, g, h)) (Exp b) (Exp b')

(Elt a, Elt b, Elt b', Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt i) => Field2 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b', c, d, e, f, g, h, i)) (Exp b) (Exp b')

Instance details

Methods

_2 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b', c, d, e, f, g, h, i)) (Exp b) (Exp b')

(Arrays a, Arrays b, Arrays c, Arrays c') => Field3 (Acc (a, b, c)) (Acc (a, b, c')) (Acc c) (Acc c')

Instance details

Methods

_3 :: Lens (Acc (a, b, c)) (Acc (a, b, c')) (Acc c) (Acc c')

(Arrays a, Arrays b, Arrays c, Arrays c', Arrays d) => Field3 (Acc (a, b, c, d)) (Acc (a, b, c', d)) (Acc c) (Acc c')

Instance details

Methods

_3 :: Lens (Acc (a, b, c, d)) (Acc (a, b, c', d)) (Acc c) (Acc c')

(Arrays a, Arrays b, Arrays c, Arrays c', Arrays d, Arrays e) => Field3 (Acc (a, b, c, d, e)) (Acc (a, b, c', d, e)) (Acc c) (Acc c')

Instance details

Methods

_3 :: Lens (Acc (a, b, c, d, e)) (Acc (a, b, c', d, e)) (Acc c) (Acc c')

(Arrays a, Arrays b, Arrays c, Arrays c', Arrays d, Arrays e, Arrays f) => Field3 (Acc (a, b, c, d, e, f)) (Acc (a, b, c', d, e, f)) (Acc c) (Acc c')

Instance details

Methods

_3 :: Lens (Acc (a, b, c, d, e, f)) (Acc (a, b, c', d, e, f)) (Acc c) (Acc c')

(Arrays a, Arrays b, Arrays c, Arrays c', Arrays d, Arrays e, Arrays f, Arrays g) => Field3 (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c', d, e, f, g)) (Acc c) (Acc c')

Instance details

Methods

_3 :: Lens (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c', d, e, f, g)) (Acc c) (Acc c')

(Arrays a, Arrays b, Arrays c, Arrays c', Arrays d, Arrays e, Arrays f, Arrays g, Arrays h) => Field3 (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c', d, e, f, g, h)) (Acc c) (Acc c')

Instance details

Methods

_3 :: Lens (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c', d, e, f, g, h)) (Acc c) (Acc c')

(Arrays a, Arrays b, Arrays c, Arrays c', Arrays d, Arrays e, Arrays f, Arrays g, Arrays h, Arrays i) => Field3 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c', d, e, f, g, h, i)) (Acc c) (Acc c')

Instance details

Methods

_3 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c', d, e, f, g, h, i)) (Acc c) (Acc c')

(Elt a, Elt b, Elt c, Elt c') => Field3 (Exp (a, b, c)) (Exp (a, b, c')) (Exp c) (Exp c')

Instance details

Methods

_3 :: Lens (Exp (a, b, c)) (Exp (a, b, c')) (Exp c) (Exp c')

(Elt a, Elt b, Elt c, Elt c', Elt d) => Field3 (Exp (a, b, c, d)) (Exp (a, b, c', d)) (Exp c) (Exp c')

Instance details

Methods

_3 :: Lens (Exp (a, b, c, d)) (Exp (a, b, c', d)) (Exp c) (Exp c')

(Elt a, Elt b, Elt c, Elt c', Elt d, Elt e) => Field3 (Exp (a, b, c, d, e)) (Exp (a, b, c', d, e)) (Exp c) (Exp c')

Instance details

Methods

_3 :: Lens (Exp (a, b, c, d, e)) (Exp (a, b, c', d, e)) (Exp c) (Exp c')

(Elt a, Elt b, Elt c, Elt c', Elt d, Elt e, Elt f) => Field3 (Exp (a, b, c, d, e, f)) (Exp (a, b, c', d, e, f)) (Exp c) (Exp c')

Instance details

Methods

_3 :: Lens (Exp (a, b, c, d, e, f)) (Exp (a, b, c', d, e, f)) (Exp c) (Exp c')

(Elt a, Elt b, Elt c, Elt c', Elt d, Elt e, Elt f, Elt g) => Field3 (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c', d, e, f, g)) (Exp c) (Exp c')

Instance details

Methods

_3 :: Lens (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c', d, e, f, g)) (Exp c) (Exp c')

(Elt a, Elt b, Elt c, Elt c', Elt d, Elt e, Elt f, Elt g, Elt h) => Field3 (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c', d, e, f, g, h)) (Exp c) (Exp c')

Instance details

Methods

_3 :: Lens (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c', d, e, f, g, h)) (Exp c) (Exp c')

(Elt a, Elt b, Elt c, Elt c', Elt d, Elt e, Elt f, Elt g, Elt h, Elt i) => Field3 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c', d, e, f, g, h, i)) (Exp c) (Exp c')

Instance details

Methods

_3 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c', d, e, f, g, h, i)) (Exp c) (Exp c')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays d') => Field4 (Acc (a, b, c, d)) (Acc (a, b, c, d')) (Acc d) (Acc d')

Instance details

Methods

_4 :: Lens (Acc (a, b, c, d)) (Acc (a, b, c, d')) (Acc d) (Acc d')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays d', Arrays e) => Field4 (Acc (a, b, c, d, e)) (Acc (a, b, c, d', e)) (Acc d) (Acc d')

Instance details

Methods

_4 :: Lens (Acc (a, b, c, d, e)) (Acc (a, b, c, d', e)) (Acc d) (Acc d')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays d', Arrays e, Arrays f) => Field4 (Acc (a, b, c, d, e, f)) (Acc (a, b, c, d', e, f)) (Acc d) (Acc d')

Instance details

Methods

_4 :: Lens (Acc (a, b, c, d, e, f)) (Acc (a, b, c, d', e, f)) (Acc d) (Acc d')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays d', Arrays e, Arrays f, Arrays g) => Field4 (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c, d', e, f, g)) (Acc d) (Acc d')

Instance details

Methods

_4 :: Lens (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c, d', e, f, g)) (Acc d) (Acc d')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays d', Arrays e, Arrays f, Arrays g, Arrays h) => Field4 (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d', e, f, g, h)) (Acc d) (Acc d')

Instance details

Methods

_4 :: Lens (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d', e, f, g, h)) (Acc d) (Acc d')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays d', Arrays e, Arrays f, Arrays g, Arrays h, Arrays i) => Field4 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d', e, f, g, h, i)) (Acc d) (Acc d')

Instance details

Methods

_4 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d', e, f, g, h, i)) (Acc d) (Acc d')

(Elt a, Elt b, Elt c, Elt d, Elt d') => Field4 (Exp (a, b, c, d)) (Exp (a, b, c, d')) (Exp d) (Exp d')

Instance details

Methods

_4 :: Lens (Exp (a, b, c, d)) (Exp (a, b, c, d')) (Exp d) (Exp d')

(Elt a, Elt b, Elt c, Elt d, Elt d', Elt e) => Field4 (Exp (a, b, c, d, e)) (Exp (a, b, c, d', e)) (Exp d) (Exp d')

Instance details

Methods

_4 :: Lens (Exp (a, b, c, d, e)) (Exp (a, b, c, d', e)) (Exp d) (Exp d')

(Elt a, Elt b, Elt c, Elt d, Elt d', Elt e, Elt f) => Field4 (Exp (a, b, c, d, e, f)) (Exp (a, b, c, d', e, f)) (Exp d) (Exp d')

Instance details

Methods

_4 :: Lens (Exp (a, b, c, d, e, f)) (Exp (a, b, c, d', e, f)) (Exp d) (Exp d')

(Elt a, Elt b, Elt c, Elt d, Elt d', Elt e, Elt f, Elt g) => Field4 (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c, d', e, f, g)) (Exp d) (Exp d')

Instance details

Methods

_4 :: Lens (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c, d', e, f, g)) (Exp d) (Exp d')

(Elt a, Elt b, Elt c, Elt d, Elt d', Elt e, Elt f, Elt g, Elt h) => Field4 (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d', e, f, g, h)) (Exp d) (Exp d')

Instance details

Methods

_4 :: Lens (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d', e, f, g, h)) (Exp d) (Exp d')

(Elt a, Elt b, Elt c, Elt d, Elt d', Elt e, Elt f, Elt g, Elt h, Elt i) => Field4 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d', e, f, g, h, i)) (Exp d) (Exp d')

Instance details

Methods

_4 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d', e, f, g, h, i)) (Exp d) (Exp d')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays e') => Field5 (Acc (a, b, c, d, e)) (Acc (a, b, c, d, e')) (Acc e) (Acc e')

Instance details

Methods

_5 :: Lens (Acc (a, b, c, d, e)) (Acc (a, b, c, d, e')) (Acc e) (Acc e')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays e', Arrays f) => Field5 (Acc (a, b, c, d, e, f)) (Acc (a, b, c, d, e', f)) (Acc e) (Acc e')

Instance details

Methods

_5 :: Lens (Acc (a, b, c, d, e, f)) (Acc (a, b, c, d, e', f)) (Acc e) (Acc e')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays e', Arrays f, Arrays g) => Field5 (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c, d, e', f, g)) (Acc e) (Acc e')

Instance details

Methods

_5 :: Lens (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c, d, e', f, g)) (Acc e) (Acc e')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays e', Arrays f, Arrays g, Arrays h) => Field5 (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d, e', f, g, h)) (Acc e) (Acc e')

Instance details

Methods

_5 :: Lens (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d, e', f, g, h)) (Acc e) (Acc e')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays e', Arrays f, Arrays g, Arrays h, Arrays i) => Field5 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e', f, g, h, i)) (Acc e) (Acc e')

Instance details

Methods

_5 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e', f, g, h, i)) (Acc e) (Acc e')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt e') => Field5 (Exp (a, b, c, d, e)) (Exp (a, b, c, d, e')) (Exp e) (Exp e')

Instance details

Methods

_5 :: Lens (Exp (a, b, c, d, e)) (Exp (a, b, c, d, e')) (Exp e) (Exp e')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt e', Elt f) => Field5 (Exp (a, b, c, d, e, f)) (Exp (a, b, c, d, e', f)) (Exp e) (Exp e')

Instance details

Methods

_5 :: Lens (Exp (a, b, c, d, e, f)) (Exp (a, b, c, d, e', f)) (Exp e) (Exp e')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt e', Elt f, Elt g) => Field5 (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c, d, e', f, g)) (Exp e) (Exp e')

Instance details

Methods

_5 :: Lens (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c, d, e', f, g)) (Exp e) (Exp e')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt e', Elt f, Elt g, Elt h) => Field5 (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d, e', f, g, h)) (Exp e) (Exp e')

Instance details

Methods

_5 :: Lens (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d, e', f, g, h)) (Exp e) (Exp e')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt e', Elt f, Elt g, Elt h, Elt i) => Field5 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e', f, g, h, i)) (Exp e) (Exp e')

Instance details

Methods

_5 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e', f, g, h, i)) (Exp e) (Exp e')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays f') => Field6 (Acc (a, b, c, d, e, f)) (Acc (a, b, c, d, e, f')) (Acc f) (Acc f')

Instance details

Methods

_6 :: Lens (Acc (a, b, c, d, e, f)) (Acc (a, b, c, d, e, f')) (Acc f) (Acc f')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays f', Arrays g) => Field6 (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c, d, e, f', g)) (Acc f) (Acc f')

Instance details

Methods

_6 :: Lens (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c, d, e, f', g)) (Acc f) (Acc f')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays f', Arrays g, Arrays h) => Field6 (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d, e, f', g, h)) (Acc f) (Acc f')

Instance details

Methods

_6 :: Lens (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d, e, f', g, h)) (Acc f) (Acc f')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays f', Arrays g, Arrays h, Arrays i) => Field6 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e, f', g, h, i)) (Acc f) (Acc f')

Instance details

Methods

_6 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e, f', g, h, i)) (Acc f) (Acc f')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt f') => Field6 (Exp (a, b, c, d, e, f)) (Exp (a, b, c, d, e, f')) (Exp f) (Exp f')

Instance details

Methods

_6 :: Lens (Exp (a, b, c, d, e, f)) (Exp (a, b, c, d, e, f')) (Exp f) (Exp f')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt f', Elt g) => Field6 (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c, d, e, f', g)) (Exp f) (Exp f')

Instance details

Methods

_6 :: Lens (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c, d, e, f', g)) (Exp f) (Exp f')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt f', Elt g, Elt h) => Field6 (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d, e, f', g, h)) (Exp f) (Exp f')

Instance details

Methods

_6 :: Lens (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d, e, f', g, h)) (Exp f) (Exp f')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt f', Elt g, Elt h, Elt i) => Field6 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e, f', g, h, i)) (Exp f) (Exp f')

Instance details

Methods

_6 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e, f', g, h, i)) (Exp f) (Exp f')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays g') => Field7 (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c, d, e, f, g')) (Acc g) (Acc g')

Instance details

Methods

_7 :: Lens (Acc (a, b, c, d, e, f, g)) (Acc (a, b, c, d, e, f, g')) (Acc g) (Acc g')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays g', Arrays h) => Field7 (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d, e, f, g', h)) (Acc g) (Acc g')

Instance details

Methods

_7 :: Lens (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d, e, f, g', h)) (Acc g) (Acc g')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays g', Arrays h, Arrays i) => Field7 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e, f, g', h, i)) (Acc g) (Acc g')

Instance details

Methods

_7 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e, f, g', h, i)) (Acc g) (Acc g')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt g') => Field7 (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c, d, e, f, g')) (Exp g) (Exp g')

Instance details

Methods

_7 :: Lens (Exp (a, b, c, d, e, f, g)) (Exp (a, b, c, d, e, f, g')) (Exp g) (Exp g')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt g', Elt h) => Field7 (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d, e, f, g', h)) (Exp g) (Exp g')

Instance details

Methods

_7 :: Lens (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d, e, f, g', h)) (Exp g) (Exp g')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt g', Elt h, Elt i) => Field7 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e, f, g', h, i)) (Exp g) (Exp g')

Instance details

Methods

_7 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e, f, g', h, i)) (Exp g) (Exp g')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays h, Arrays h') => Field8 (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d, e, f, g, h')) (Acc h) (Acc h')

Instance details

Methods

_8 :: Lens (Acc (a, b, c, d, e, f, g, h)) (Acc (a, b, c, d, e, f, g, h')) (Acc h) (Acc h')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays h, Arrays h', Arrays i) => Field8 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e, f, g, h', i)) (Acc h) (Acc h')

Instance details

Methods

_8 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e, f, g, h', i)) (Acc h) (Acc h')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt h') => Field8 (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d, e, f, g, h')) (Exp h) (Exp h')

Instance details

Methods

_8 :: Lens (Exp (a, b, c, d, e, f, g, h)) (Exp (a, b, c, d, e, f, g, h')) (Exp h) (Exp h')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt h', Elt i) => Field8 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e, f, g, h', i)) (Exp h) (Exp h')

Instance details

Methods

_8 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e, f, g, h', i)) (Exp h) (Exp h')

(Arrays a, Arrays b, Arrays c, Arrays d, Arrays e, Arrays f, Arrays g, Arrays h, Arrays i, Arrays i') => Field9 (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e, f, g, h, i')) (Acc i) (Acc i')

Instance details

Methods

_9 :: Lens (Acc (a, b, c, d, e, f, g, h, i)) (Acc (a, b, c, d, e, f, g, h, i')) (Acc i) (Acc i')

(Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt i, Elt i') => Field9 (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e, f, g, h, i')) (Exp i) (Exp i')

Instance details

Methods

_9 :: Lens (Exp (a, b, c, d, e, f, g, h, i)) (Exp (a, b, c, d, e, f, g, h, i')) (Exp i) (Exp i')