Data.Array.Accelerate.Control.Lens.Shape

Description

Documentation

module Control.Lens.Tuple

Orphan instances

(Slice sh, Elt a, Elt a') => Field1 (Exp ((:.) sh a)) (Exp ((:.) sh a')) (Exp a) (Exp a')

Methods

_1 :: Lens (Exp (sh :. a)) (Exp (sh :. a')) (Exp a) (Exp a')

(Slice sh, Elt a, Elt b, Elt b', Slice ((:.) sh b), Slice ((:.) sh b')) => Field2 (Exp ((:.) ((:.) sh b) a)) (Exp ((:.) ((:.) sh b') a)) (Exp b) (Exp b')

Methods

_2 :: Lens (Exp ((sh :. b) :. a)) (Exp ((sh :. b') :. a)) (Exp b) (Exp b')

(Slice sh, Elt a, Elt b, Elt c, Elt c', Slice ((:.) sh c), Slice ((:.) ((:.) sh c) b), Slice ((:.) sh c'), Slice ((:.) ((:.) sh c') b)) => Field3 (Exp ((:.) ((:.) ((:.) sh c) b) a)) (Exp ((:.) ((:.) ((:.) sh c') b) a)) (Exp c) (Exp c')

Methods

_3 :: Lens (Exp (((sh :. c) :. b) :. a)) (Exp (((sh :. c') :. b) :. a)) (Exp c) (Exp c')

(Slice sh, Elt a, Elt b, Elt c, Elt d, Elt d', Slice ((:.) sh d), Slice ((:.) ((:.) sh d) c), Slice ((:.) ((:.) ((:.) sh d) c) b), Slice ((:.) sh d'), Slice ((:.) ((:.) sh d') c), Slice ((:.) ((:.) ((:.) sh d') c) b)) => Field4 (Exp ((:.) ((:.) ((:.) ((:.) sh d) c) b) a)) (Exp ((:.) ((:.) ((:.) ((:.) sh d') c) b) a)) (Exp d) (Exp d')

Methods

_4 :: Lens (Exp ((((sh :. d) :. c) :. b) :. a)) (Exp ((((sh :. d') :. c) :. b) :. a)) (Exp d) (Exp d')

(Slice sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt e', Slice ((:.) sh e), Slice ((:.) ((:.) sh e) d), Slice ((:.) ((:.) ((:.) sh e) d) c), Slice ((:.) ((:.) ((:.) ((:.) sh e) d) c) b), Slice ((:.) sh e'), Slice ((:.) ((:.) sh e') d), Slice ((:.) ((:.) ((:.) sh e') d) c), Slice ((:.) ((:.) ((:.) ((:.) sh e') d) c) b)) => Field5 (Exp ((:.) ((:.) ((:.) ((:.) ((:.) sh e) d) c) b) a)) (Exp ((:.) ((:.) ((:.) ((:.) ((:.) sh e') d) c) b) a)) (Exp e) (Exp e')

Methods

_5 :: Lens (Exp (((((sh :. e) :. d) :. c) :. b) :. a)) (Exp (((((sh :. e') :. d) :. c) :. b) :. a)) (Exp e) (Exp e')

(Slice sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt f', Slice ((:.) sh f), Slice ((:.) ((:.) sh f) e), Slice ((:.) ((:.) ((:.) sh f) e) d), Slice ((:.) ((:.) ((:.) ((:.) sh f) e) d) c), Slice ((:.) ((:.) ((:.) ((:.) ((:.) sh f) e) d) c) b), Slice ((:.) sh f'), Slice ((:.) ((:.) sh f') e), Slice ((:.) ((:.) ((:.) sh f') e) d), Slice ((:.) ((:.) ((:.) ((:.) sh f') e) d) c), Slice ((:.) ((:.) ((:.) ((:.) ((:.) sh f') e) d) c) b)) => Field6 (Exp ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh f) e) d) c) b) a)) (Exp ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh f') e) d) c) b) a)) (Exp f) (Exp f')

Methods

_6 :: Lens (Exp ((((((sh :. f) :. e) :. d) :. c) :. b) :. a)) (Exp ((((((sh :. f') :. e) :. d) :. c) :. b) :. a)) (Exp f) (Exp f')

(Slice sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt g', Slice ((:.) sh g), Slice ((:.) ((:.) sh g) f), Slice ((:.) ((:.) ((:.) sh g) f) e), Slice ((:.) ((:.) ((:.) ((:.) sh g) f) e) d), Slice ((:.) ((:.) ((:.) ((:.) ((:.) sh g) f) e) d) c), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh g) f) e) d) c) b), Slice ((:.) sh g'), Slice ((:.) ((:.) sh g') f), Slice ((:.) ((:.) ((:.) sh g') f) e), Slice ((:.) ((:.) ((:.) ((:.) sh g') f) e) d), Slice ((:.) ((:.) ((:.) ((:.) ((:.) sh g') f) e) d) c), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh g') f) e) d) c) b)) => Field7 (Exp ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh g) f) e) d) c) b) a)) (Exp ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh g') f) e) d) c) b) a)) (Exp g) (Exp g')

Methods

_7 :: Lens (Exp (((((((sh :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp (((((((sh :. g') :. f) :. e) :. d) :. c) :. b) :. a)) (Exp g) (Exp g')

(Slice sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt h', Slice ((:.) sh h), Slice ((:.) ((:.) sh h) g), Slice ((:.) ((:.) ((:.) sh h) g) f), Slice ((:.) ((:.) ((:.) ((:.) sh h) g) f) e), Slice ((:.) ((:.) ((:.) ((:.) ((:.) sh h) g) f) e) d), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh h) g) f) e) d) c), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh h) g) f) e) d) c) b), Slice ((:.) sh h'), Slice ((:.) ((:.) sh h') g), Slice ((:.) ((:.) ((:.) sh h') g) f), Slice ((:.) ((:.) ((:.) ((:.) sh h') g) f) e), Slice ((:.) ((:.) ((:.) ((:.) ((:.) sh h') g) f) e) d), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh h') g) f) e) d) c), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh h') g) f) e) d) c) b)) => Field8 (Exp ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh h) g) f) e) d) c) b) a)) (Exp ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh h') g) f) e) d) c) b) a)) (Exp h) (Exp h')

Methods

_8 :: Lens (Exp ((((((((sh :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp ((((((((sh :. h') :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp h) (Exp h')

(Slice sh, Elt a, Elt b, Elt c, Elt d, Elt e, Elt f, Elt g, Elt h, Elt i, Elt i', Slice ((:.) sh i), Slice ((:.) ((:.) sh i) h), Slice ((:.) ((:.) ((:.) sh i) h) g), Slice ((:.) ((:.) ((:.) ((:.) sh i) h) g) f), Slice ((:.) ((:.) ((:.) ((:.) ((:.) sh i) h) g) f) e), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh i) h) g) f) e) d), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh i) h) g) f) e) d) c), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh i) h) g) f) e) d) c) b), Slice ((:.) sh i'), Slice ((:.) ((:.) sh i') h), Slice ((:.) ((:.) ((:.) sh i') h) g), Slice ((:.) ((:.) ((:.) ((:.) sh i') h) g) f), Slice ((:.) ((:.) ((:.) ((:.) ((:.) sh i') h) g) f) e), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh i') h) g) f) e) d), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh i') h) g) f) e) d) c), Slice ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh i') h) g) f) e) d) c) b)) => Field9 (Exp ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh i) h) g) f) e) d) c) b) a)) (Exp ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) ((:.) sh i') h) g) f) e) d) c) b) a)) (Exp i) (Exp i')

Methods

_9 :: Lens (Exp (((((((((sh :. i) :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp (((((((((sh :. i') :. h) :. g) :. f) :. e) :. d) :. c) :. b) :. a)) (Exp i) (Exp i')